requirements.txt
boto3==1.24.7
ipykernel==6.17.1
pandas==1.4.1
opencv-contrib-python==3.4.17.63
matplotlib-inline==0.1.6
matplotlib==3.6.2
pyyaml==6.0